requirements.txt
Django
django-allauth
django-widget-tweaks
Pillow
apscheduler
psycopg2